feat(header): implement compact mode for header
