+++ client/package.json
@@ -1,7 +1,7 @@
 {
   "name": "app-react",
   "private": true,
-  "version": "0.8.1",
+  "version": "0.8.2",
   "type": "module",
   "scripts": {
     "dev": "vite",
@@ -12,6 +12,7 @@
   "dependencies": {
     "@reduxjs/toolkit": "^2.3.0",
     "axios": "^1.7.7",
+    "classnames": "^2.5.1",
     "framer-motion": "^11.11.10",
     "hamburgers": "^1.2.1",
     "leaflet": "^1.9.4",
